Remove code for has_many and belongs_to proxies
The whole proxy idea was a bad idea - especially to do for all repositories.
The whole point of this library is to manually specify the SQL, and the proxies
violated that principle. The proxies were also terrible performance-wise.
After looking at how ActiveRecord and Sequel do eager loading of associations,
we have a better way to handle has_many and belongs_to relationships.