Require `sequel` when requiring `preserves`
This ensures that everything you need will be provided with a single
`require "preserves"` call.  It also means that `rake console` will include
Preserves as well as Sequel.

NOTE: It doesn't seem possible to require Sequel without Sequel::Model.
That's disappointing that we have to pull in all that unused code.